move caching reader and text extraction to util function
