+++ pkg/upstream/replicated.go
@@ -412,12 +412,12 @@ func downloadReplicatedApp(replicatedUpstream *replicatedapp.ReplicatedUpstream,
 		replicatedChartNames = strings.Split(replicatedChartNamesStr, ",")
 	}
 
-	cachingBodyReader := newCachingReader(getResp.Body, 4096)
+	cachingBodyReader := util.NewCachingReader(getResp.Body, 4096)
 	gzf, err := gzip.NewReader(cachingBodyReader)
 	if err != nil {
 		io.ReadAll(cachingBodyReader)
 		cachedContent := cachingBodyReader.GetCachedBytes()
-		readableText := extractReadableText(cachedContent)
+		readableText := util.ExtractReadableText(cachedContent)
 		if len(readableText) > 0 {
 			return nil, errors.Wrapf(err, "failed to create gzip reader, ended with %q", readableText)
 		}
@@ -450,7 +450,7 @@ func downloadReplicatedApp(replicatedUpstream *replicatedapp.ReplicatedUpstream,
 		if err != nil {
 			io.ReadAll(cachingBodyReader)
 			cachedContent := cachingBodyReader.GetCachedBytes()
-			readableText := extractReadableText(cachedContent)
+			readableText := util.ExtractReadableText(cachedContent)
 			if len(readableText) > 0 {
 				return nil, errors.Wrapf(err, "failed to get next file from reader, ended with %q", readableText)
 			}
@@ -467,7 +467,7 @@ func downloadReplicatedApp(replicatedUpstream *replicatedapp.ReplicatedUpstream,
 			if err != nil {
 				io.ReadAll(cachingBodyReader)
 				cachedContent := cachingBodyReader.GetCachedBytes()
-				readableText := extractReadableText(cachedContent)
+				readableText := util.ExtractReadableText(cachedContent)
 				if len(readableText) > 0 {
 					return nil, errors.Wrapf(err, "failed to read file from tar, ended with %q", readableText)
 				}
@@ -481,46 +481,6 @@ func downloadReplicatedApp(replicatedUpstream *replicatedapp.ReplicatedUpstream,
 	}
 
 	return &release, nil
-}
-
-// extractReadableText takes a byte slice and returns a string containing only the readable ASCII text
-// from the original data. This is useful for logging binary data in a human-readable format.
-func extractReadableText(data []byte) string {
-	if len(data) == 0 {
-		return ""
-	}
-
-	var result strings.Builder
-
-	// Find runs of printable characters that are at least 5 characters long
-	var currentRun strings.Builder
-	for _, b := range data {
-		// Only include printable ASCII characters (32-126) and newlines/tabs
-		if (b >= 32 && b <= 126) || b == '\n' || b == '\t' || b == '\r' {
-			currentRun.WriteByte(b)
-		} else {
-			// Non-printable character encountered, check the current run
-			if currentRun.Len() >= 5 {
-				// If the run is long enough, add it to the result with spaces between runs
-				if result.Len() > 0 {
-					result.WriteString(" ... ")
-				}
-				result.WriteString(currentRun.String())
-			}
-			// Reset the current run
-			currentRun.Reset()
-		}
-	}
-
-	// Check the last run
-	if currentRun.Len() >= 5 {
-		if result.Len() > 0 {
-			result.WriteString(" ... ")
-		}
-		result.WriteString(currentRun.String())
-	}
-
-	return result.String()
 }
 
 func listPendingChannelReleases(license *kotsv1beta1.License, lastUpdateCheckAt *time.Time, currentCursor replicatedapp.ReplicatedCursor, channelChanged bool, sortOrder string, reportingInfo *reportingtypes.ReportingInfo, selectedChannelID string) ([]ChannelRelease, *time.Time, error) {
